test(api): isolate editor preflight regression
